Copy with range to server #2051 - PR
